Rename eos to arista_eos to be consistent with ohai. Signed-off-by: Jere Julian <jere@arista.com>
