+++ lib/train/extras/os_detect_arista_eos.rb
